Migrated to Styled Components and Configured Husky

+++ pages/api/post/[id].ts
@@ -1,18 +1,14 @@
-// pages/api/post/[id].ts
-
+import { NextApiRequest, NextApiResponse } from 'next';
 import prisma from '../../../lib/prisma';
 
-// DELETE /api/post/:id
-export default async function handle(req, res) {
-  const postId = req.query.id;
-  if (req.method === 'DELETE') {
-    const post = await prisma.post.delete({
-      where: { id: postId },
-    });
-    res.json(post);
-  } else {
-    throw new Error(
-      `The HTTP ${req.method} method is not supported at this route.`,
-    );
-  }
+export default async function handle(req: NextApiRequest, res: NextApiResponse) {
+    const { id } = req.query;
+    if (req.method === 'DELETE' && typeof id === 'string') {
+        const post = await prisma.post.delete({
+            where: { id },
+        });
+        res.json(post);
+    } else {
+        throw new Error(`The HTTP ${req.method} method is not supported at this route.`);
+    }
 }
